use password for edit and delete operations

+++ src/controllers/genresController.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from "express";
 
-import type { GameData } from "../types/types.js";
+import type { AdminPassword, GameData } from "../types/types.js";
 
 import {
   deleteGenre,
@@ -12,6 +12,15 @@ import {
 } from "../db/queries.js";
 
 async function changeGenreName(req: Request, res: Response) {
+  const { adminPassword } = req.body as AdminPassword;
+  if (!adminPassword || adminPassword !== process.env.ADMIN_PASSWORD) {
+    return res
+      .status(401)
+      .send(
+        "<script>alert('Invalid admin password'); window.history.back();</script>",
+      );
+  }
+
   const genreId = req.params.id;
   const newGenreName = (req.body as GameData).name;
   await updateGenreName(genreId, newGenreName);
@@ -19,6 +28,15 @@ async function changeGenreName(req: Request, res: Response) {
 }
 
 async function removeGenre(req: Request, res: Response) {
+  const { adminPassword } = req.body as AdminPassword;
+  if (!adminPassword || adminPassword !== process.env.ADMIN_PASSWORD) {
+    return res
+      .status(401)
+      .send(
+        "<script>alert('Invalid admin password'); window.history.back();</script>",
+      );
+  }
+
   const genreId = req.params.id;
   await deleteGenre(genreId);
   res.redirect("/genres");
